requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
-e .
dawdreamer
einops
fastapi
hydra-core
hydra-colorlog
lightning
omegaconf
pedalboard
python-dotenv
rich
soundfile
timm==0.4.12
torch
torchaudio
torchmetrics
torchopenl3
tqdm
wandb